Completed endianness change.

+++ core/src/main/java/com/yonathan/chess/core/board/AbstractBitboardMoveGenerator.java
@@ -224,7 +224,7 @@ public abstract class AbstractBitboardMoveGenerator extends AbstractBitboardMagi
     }
 
 
-    public int getEPSquare() {
+    public long getEPSquare() {
         return ePSquare;
     }
